requirements.txt
numpy>=1.24
sentence-transformers>=2.6
transformers>=4.38
torch>=2.1
pymupdf>=1.23
python-docx>=1.1
requests>=2.31
pysbd>=0.3.4
rich>=13.7
pyyaml>=6.0
beautifulsoup4>=4.12
lxml>=5.2
